RUB-302: split sig queue tests by module
Move the large sig_queue test module into focused cfg(test) sibling modules.

No runtime logic, public API, Cargo, fixture, or conformance changes.
